chore: move core into its own mod
